+++ packages/abell/defaults/vite.config.js
@@ -1,3 +1,3 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const { defineConfig } = require('../index');
+const { defineConfig } = require('../src/index');
 module.exports = defineConfig({});
