+++ packages/abell/package.json
@@ -7,6 +7,7 @@
   "types": "./dist/index.d.ts",
   "files": [
     "dist",
+    "defaults",
     "types.d.ts"
   ],
   "scripts": {
