+++ packages/abell/scripts/post-build.js
@@ -1,63 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 
-/**
- * Recursively creates the path
- * @param {String} pathToCreate path that you want to create
- */
-function createPathIfAbsent(pathToCreate) {
-  // prettier-ignore
-  pathToCreate
-    .split(path.sep)
-    .reduce((prevPath, folder) => {
-      const currentPath = path.join(prevPath, folder, path.sep);
-      if (!fs.existsSync(currentPath)) {
-        fs.mkdirSync(currentPath);
-      }
-      return currentPath;
-    }, '');
-}
-
-/**
- *
- * @param {String} from - Path to copy from
- * @param {String} to - Path to copy to
- * @param {Array} ignore - files/directories to ignore
- * @param {Boolean} ignoreEmptyDirs - Ignore empty directories while copying
- * @return {void}
- */
-function copyFolderSync(from, to, ignore = [], ignoreEmptyDirs = true) {
-  if (ignore.includes(from)) {
-    return;
-  }
-  const fromDirectories = fs.readdirSync(from);
-
-  createPathIfAbsent(to);
-  fromDirectories.forEach((element) => {
-    const fromElement = path.join(from, element);
-    const toElement = path.join(to, element);
-    if (fs.lstatSync(fromElement).isFile()) {
-      if (!ignore.includes(fromElement)) {
-        fs.copyFileSync(fromElement, toElement);
-      }
-    } else {
-      copyFolderSync(fromElement, toElement, ignore);
-      if (fs.existsSync(toElement) && ignoreEmptyDirs) {
-        try {
-          fs.rmdirSync(toElement);
-        } catch (err) {
-          if (err.code !== 'ENOTEMPTY') throw err;
-        }
-      }
-    }
-  });
-}
-
 const DIST = path.join(__dirname, '..', 'dist');
-const from = path.join(__dirname, '..', 'src', 'defaults');
-const to = path.join(__dirname, '..', 'dist', 'defaults');
-copyFolderSync(from, to);
-console.log('> Copied defaults');
 
 try {
   const fd = fs.openSync(path.join(DIST, 'bin.js'), 'r');
