+++ packages/abell/src/index.ts
@@ -1,3 +1,8 @@
-export { compile } from './compiler';
-export { vitePluginAbell } from './utils/vite-plugin-abell';
-export { defineConfig, findAbellFileFromURL } from './utils/general-utils';
+// @TODO: export when it is ready for external use
+// export { compile } from './vite-plugin-abell/compiler';
+export { vitePluginAbell } from './vite-plugin-abell';
+export {
+  defineConfig,
+  findAbellFileFromURL,
+  AbellViteConfig
+} from './utils/api';
