+++ packages/abell/src/utils/__tests__/test-utils.ts
@@ -1,0 +1,5 @@
+import path from 'path';
+
+export const BASE_PATH = path.join(__dirname, 'test-files');
+export const prefix = (fileName: string, baseOverride?: string): string =>
+  path.join(baseOverride ?? BASE_PATH, fileName);
