+++ packages/abell/src/vite-plugin-abell/index.ts
@@ -1,0 +1,1 @@
+export * from './vite-plugin-abell';
