--- packages/abell/src/vite.ts
@@ -1,5 +1,0 @@
-/**
- * Re-export of vite to avoid versioning issues with abell
- */
-export * from 'vite';
-export { default } from 'vite';
