+++ packages/abell/tsconfig.json
@@ -6,5 +6,5 @@
     "rootDir": "src",
   },
   "include": ["src"],
-  "exclude": ["node_modules", "dist", "scripts","src/defaults"]
+  "exclude": ["node_modules", "dist", "scripts","defaults"]
 }
